+++ continuousintegration/orchestration/create_instance.sh
@@ -46,9 +46,6 @@ runoracleunittests=0
 # By default doesn't prepare the images with the previous schema version
 updatedatabasetest=0
 
-# By default doesn't run the tests for external tape formats
-runexternaltapetests=0
-
 # Create an instance for
 systest_only=0
 
@@ -65,7 +62,6 @@ Options:
   -a    additional kubernetes resources added to the kubernetes namespace
   -U    Run database unit test only
   -u    Prepare the pods to run the liquibase test
-  -T    Execute tests for external tape formats
   -Q    Create the cluster using the last ctageneric image from main
 EOF
 exit 1
@@ -115,9 +111,6 @@ while getopts "n:o:d:e:a:p:b:i:B:E:SDOUumTQ" o; do
             ;;
         U)
             runoracleunittests=1
-            ;;
-        T)
-            runexternaltapetests=1
             ;;
         u)
             updatedatabasetest=1
@@ -482,22 +475,5 @@ fi
 
 echo "Instance ${instance} successfully created:"
 kubectl --namespace=${instance} get pod ${KUBECTL_DEPRECATED_SHOWALL}
-
-if [ $runexternaltapetests == 1 ] ; then
-  echo "Running database unit-tests"
-  ./tests/external_tapes_test.sh -n ${instance} -P ${poddir}
-
-  kubectl --namespace=${instance} logs externaltapetests
-
-  # database unit-tests went wrong => exit now with error
-  if $(kubectl --namespace=${instance} get pod externaltapetests ${KUBECTL_DEPRECATED_SHOWALL} -o json | jq -r .status.phase | egrep -q 'Failed'); then
-    echo "externaltapetests pod in Failed status here are its last log lines:"
-    kubectl --namespace=${instance} logs externaltapetests --tail 10
-    die "ERROR: externaltapetests pod in Error state. Initialization failed."
-  fi
-
-  # database unit-tests were successful => exit now with success
-  exit 0
-fi
 
 exit 0