+++ continuousintegration/orchestration/tests/read_osm_tape.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # @project      The CERN Tape Archive (CTA)
-# @copyright    Copyright © 2022 CERN
+# @copyright    Copyright © 2024 CERN
 # @license      This program is free software, distributed under the terms of the GNU General Public
 #               Licence version 3 (GPL Version 3), copied verbatim in the file "COPYING". You can
 #               redistribute it and/or modify it under the terms of the GPL Version 3, or (at your
@@ -15,14 +15,30 @@
 #               granted to it by virtue of its status as an Intergovernmental Organization or
 #               submit itself to any jurisdiction.
 
-. /opt/run/bin/init_pod.sh
 
-set -e
 
-yum-config-manager --enable cta-artifacts
-yum-config-manager --enable ceph
+# Download OSM sample tape
+yum -y install git git-lfs
+git lfs install --skip-repo
+git clone https://gitlab.desy.de/mwai.karimi/osm-mhvtl.git /osm-mhvtl
 
-# Install missing RPMs
-yum -y install mt-st lsscsi sg3_utils cta-taped cta-tape-label cta-debuginfo ceph-common cta-systemtests
+device="$1"
+# Load tape in a tapedrive
+mtx -f /dev/smc status
+mtx -f /dev/smc load 1 0
+mtx -f /dev/smc status
 
-2>&1 cta-osmReaderTest ${DEVICE_NAME} ${DRIVE_DEVICE}
+# Get the device status where the tape is loaded and rewind it.
+mt -f ${device} status
+mt -f ${device} rewind
+
+# The header files have 4 more bytes in the git file
+truncate -s -4 /osm-mhvtl/L08033/L1
+touch /osm-tape.img
+dd if=/osm-mhvtl/L08033/L1 of=/osm-tape.img bs=32768
+dd if=/osm-mhvtl/L08033/L2 of=/osm-tape.img bs=32768 seek=1
+dd if=/osm-tape.img of=$device bs=32768 count=2
+dd if=/osm-mhvtl/L08033/file1 of=$device bs=262144 count=202
+dd if=/osm-mhvtl/L08033/file2 of=$device bs=262144 count=202
+
+mt -f \$device rewind